// src/models/enums/AnthropometryMethod.ts
export enum AnthropometryMethod {
    // Adultos
    DURNIN_WOMERSLEY = "DURNIN_WOMERSLEY",
    FAULKNER = "FAULKNER",
    GUEDES = "GUEDES",
    JACKSON_POLLOCK_3 = "JACKSON_POLLOCK_3",
    JACKSON_POLLOCK_7 = "JACKSON_POLLOCK_7",
    JACKSON_POLLOCK_WARD_3 = "JACKSON_POLLOCK_WARD_3",
    JACKSON_POLLOCK_WARD_7 = "JACKSON_POLLOCK_WARD_7",
    PETROSKI = "PETROSKI",
  
    // Pediatria
    WHO_BAZ = "WHO_BAZ",   // IMC-para-idade (5–19a)
    WHO_HAZ = "WHO_HAZ",   // Estatura-para-idade
    WHO_WHZ = "WHO_WHZ",   // Peso-para-estatura (0–5a)
    SLAUGHTER = "SLAUGHTER"// %G por TR+SB (8–18a)
  }